Move table and filter fetching into their components
Moves the code for fetching table info and filters from `TablePage`
into `Table` and `TableFilters`.
The goal here is to prevent the fetching code from rerendering
unrelate components in `TablePage`.
